patch: Refactor playlist definition to make URI optional
Because the system will first support HTTP push based data path for sending new playlist into the system, the playlist received will not have a source URI as it is sent to the system. In addition, the video segments listed can not be found relative to that playlist.

So we remove URI field from the playlist data structure. The playlist parser will derive the URI of each video segments in a playlist based on a base URI prefix supplied to the parser. Now the URI of the video segments is decoupled from the URI of the playlist.