patch[persistence]: Rework DB Session Management
Rework the DB session management. Remove transactions from individual API calls of `persistenceManagerImpl`, and Change `PersistenceManager` into a transient object. Each time a `PersistenceManager` is created, it is given one SQL transaction to perform all the SQL operations with. `PersistenceManager.Close()` must be called before exiting this scope to commit or rollback all the SQL operations of that transaction.