feat: Refactor Command And Control System
Refactor the command and control system between the control and edge
nodes. The RPC from the control node to the edge node is still based
on PubSub, but the RPC from the edge node to the control node will use
a REST channel

This REST based system handles the less critical RPC calls from the
edge node to the control node.
* Video source status updates
* Active recording session state sync
* Video forwarding state sync

The reporting of the recording session segments is still using
PubSub, as that information is critical.